+++ mex/extractors/artificial/provider.py
@@ -11,7 +11,7 @@ from faker.providers.python import Provider as PythonFakerProvider
 from pydantic.fields import FieldInfo
 
 from mex.common.identity import Identity
-from mex.common.models import ExtractedData
+from mex.common.models import AnyExtractedModel
 from mex.common.types import (
     TEMPORAL_ENTITY_FORMATS_BY_PRECISION,
     UTC,
@@ -98,21 +98,24 @@ class BuilderProvider(PythonFakerProvider):
         elif issubclass(inner_type, int):
             factory = self.generator.random_int
         else:
-            raise RuntimeError(f"Cannot create fake data for {field}")
+            msg = f"Cannot create fake data for {field}"
+            raise RuntimeError(msg)
         return [factory() for _ in range(self.pyint(*self.min_max_for_field(field)))]
 
-    def extracted_data(self, model: type[ExtractedData]) -> list[ExtractedData]:
-        """Get a list of extracted data instances for the given model class."""
+    def extracted_items(
+        self, model: type[AnyExtractedModel]
+    ) -> list[AnyExtractedModel]:
+        """Get a list of extracted items for the given model class."""
         models = []
         for identity in cast(list[Identity], self.generator.identities(model)):
             # manually set identity related fields
-            payload: dict[str, Any] = dict(
-                identifier=identity.identifier,
-                hadPrimarySource=identity.hadPrimarySource,
-                identifierInPrimarySource=identity.identifierInPrimarySource,
-                stableTargetId=identity.stableTargetId,
-                entityType=model.__name__,
-            )
+            payload: dict[str, Any] = {
+                "identifier": identity.identifier,
+                "hadPrimarySource": identity.hadPrimarySource,
+                "identifierInPrimarySource": identity.identifierInPrimarySource,
+                "stableTargetId": identity.stableTargetId,
+                "entityType": model.__name__,
+            }
             # dynamically populate all other fields
             for name, field in model.model_fields.items():
                 if name not in payload:
@@ -129,7 +132,7 @@ class IdentityProvider(BaseFakerProvider):
         super().__init__(factory)
         self._identities = identities
 
-    def identities(self, model: type[ExtractedData]) -> list[Identity]:
+    def identities(self, model: type[AnyExtractedModel]) -> list[Identity]:
         """Return a list of identities for the given model class."""
         return self._identities[model.__name__.removeprefix("Extracted")]
 
@@ -174,7 +177,7 @@ class TemporalEntityProvider(PythonFakerProvider):
         """Return a custom temporal entity with random date, time and precision."""
         return TemporalEntity(
             datetime.fromtimestamp(
-                self.pyint(int(8e8), int(datetime.now().timestamp())), tz=UTC
+                self.pyint(int(8e8), int(datetime.now(tz=UTC).timestamp())), tz=UTC
             ).strftime(
                 TEMPORAL_ENTITY_FORMATS_BY_PRECISION[
                     self.random_element(allowed_precision_levels)