+++ mex/extractors/ff_projects/extract.py
@@ -11,28 +11,15 @@ from mex.common.ldap.models.person import LDAPPersonWithQuery
 from mex.common.ldap.transform import analyse_person_string
 from mex.common.logging import watch
 from mex.common.types import (
+    MergedOrganizationIdentifier,
     TemporalEntity,
     TemporalEntityPrecision,
 )
-from mex.common.wikidata.extract import search_organization_by_label
-from mex.common.wikidata.models.organization import WikidataOrganization
 from mex.extractors.ff_projects.models.source import FFProjectsSource
 from mex.extractors.settings import Settings
-
-ORGANIZATIONS_BY_ABBREVIATIONS = {
-    "BMG": "Federal Ministry of Health of Germany",
-    "EC": "European Commission",
-    "UBA": "Federal Environment Agency",
-    "BBK": "Bundesamt für Bevölkerungsschutz und Katastrophenhilfe",
-    "DGE": "Deutsche Gesellschaft für Ernährung",
-    "FLI": "Friedrich Loeffler Institute",
-    "BMWK (ehem. BMWi)": "Federal Ministry for Economic Affairs and Climate Action",
-    "BMWK": "Federal Ministry for Economic Affairs and Climate Action",
-    "BMWiW": "Federal Ministry for Economic Affairs and Climate Action",
-    "IBB": "Investitionsbank Berlin",
-    "Sepsis-Stiftung": "Sepsis-Stiftung",
-    "DG-Sante": "Directorate-General for Health and Food Safety",
-}
+from mex.extractors.wikidata.helpers import (
+    get_wikidata_extracted_organization_id_by_name,
+)
 
 
 @watch
@@ -213,22 +200,26 @@ def extract_ff_project_authors(
 
 def extract_ff_projects_organizations(
     ff_projects_sources: Iterable[FFProjectsSource],
-) -> dict[str, WikidataOrganization]:
+) -> dict[str, MergedOrganizationIdentifier]:
     """Search and extract organization from wikidata.
 
     Args:
         ff_projects_sources: Iterable of ff-project sources
 
     Returns:
-        Dict with organization label and WikidataOrganization
+        Dict with organization label and WikidataOrganization ID
     """
-    organizations = {
-        source.zuwendungs_oder_auftraggeber: org
+    return {
+        zuwendungs_oder_auftraggeber: org_id
         for source in ff_projects_sources
         if source.zuwendungs_oder_auftraggeber
-        and (org := search_organization_by_label(source.zuwendungs_oder_auftraggeber))
+        and source.zuwendungs_oder_auftraggeber != "Sonderforschung"
+        for zuwendungs_oder_auftraggeber in source.zuwendungs_oder_auftraggeber.split(
+            "/"
+        )
+        if (
+            org_id := get_wikidata_extracted_organization_id_by_name(
+                zuwendungs_oder_auftraggeber
+            )
+        )
     }
-    for abbreviation, organization in ORGANIZATIONS_BY_ABBREVIATIONS.items():
-        if wiki_org := search_organization_by_label(organization):
-            organizations[abbreviation] = wiki_org
-    return organizations