+++ mex/extractors/ff_projects/transform.py
@@ -39,6 +39,7 @@ def transform_ff_projects_source_to_extracted_activity(
         responsible_unit = [
             unit_stable_target_id_by_synonym[oe]
             for oe in rki_oe.replace("/", ",").split(",")
+            if oe in unit_stable_target_id_by_synonym
         ]
     else:
         raise MExError("missing unit should have been filtered out")
@@ -52,14 +53,15 @@ def transform_ff_projects_source_to_extracted_activity(
     orgs = ff_projects_source.zuwendungs_oder_auftraggeber.replace("/", ",").split(",")
     funder_or_commissioner: list[MergedOrganizationIdentifier] = []
     for org in orgs:
-        if org in ["Sonderforschung", "AA"]:
+        if (
+            org
+            in ff_projects_activity.funderOrCommissioner[0].mappingRules[1].forValues
+        ):
             continue
         if org in organization_stable_target_id_by_synonyms:
             funder_or_commissioner.append(
                 organization_stable_target_id_by_synonyms[org]
             )
-        elif sti := organization_stable_target_id_by_synonyms.get(org):
-            funder_or_commissioner.append(sti)
     activity_type = []
     if (
         ff_projects_source.rki_az